+++ packages/app/src/renderer/App.vue.ts
@@ -1,0 +1,9 @@
+// import { defineComponent } from 'vue'
+import HelloWorld from './components/HelloWorld.vue'
+
+export default defineComponent/* <- try auto import here */({
+  name: 'App',
+  components: {
+    HelloWorld
+  }
+})
