+++ packages/app/src/renderer/components/HelloWorld.vue.ts
@@ -1,0 +1,14 @@
+import { ref, defineComponent } from 'vue'
+export default defineComponent({
+  name: 'HelloWorld',
+  props: {
+    msg: {
+      type: String,
+      required: true
+    }
+  },
+  setup: () => {
+    const count = ref(0)
+    return { count }
+  }
+})
